Implemented caching logic for company financials

+++ src/main/java/org/application/mars/MarketData/dtos/OHLCVBarDTO.java
@@ -2,12 +2,14 @@ package org.application.mars.MarketData.dtos;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class OHLCVBarDTO {
     String time;
-    double open;
-    double high;
-    double low;
-    double close;
+    BigDecimal open;
+    BigDecimal high;
+    BigDecimal low;
+    BigDecimal close;
     long volume;
 }
