+++ src/main/java/org/application/mars/MarketData/models/Massive/AggregateBars/CustomBars.java
@@ -3,14 +3,16 @@ package org.application.mars.MarketData.models.Massive.AggregateBars;
 import com.fasterxml.jackson.annotation.JsonIgnoreProperties;
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 @JsonIgnoreProperties(ignoreUnknown = true)
 public class CustomBars {
-    private Long c;
-    private Long h;
-    private Long l;
+    private BigDecimal c;
+    private BigDecimal h;
+    private BigDecimal l;
     private Long n;
-    private Long o;
+    private BigDecimal o;
     private Long otc;
     private Long t;
     private Long v;
